move TestSyncer
