+++ pilot/pkg/config/kube/ingress/status_test.go
@@ -15,20 +15,14 @@
 package ingress
 
 import (
-	"os"
 	"testing"
-	"time"
 
-	v1 "k8s.io/api/core/v1"
 	extensions "k8s.io/api/extensions/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/fake"
 	"k8s.io/ingress/core/pkg/ingress/annotations/class"
 
 	meshconfig "istio.io/api/mesh/v1alpha1"
-	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/serviceregistry/kube"
-	"istio.io/istio/pkg/log"
 )
 
 func makeAnnotatedIngress(annotation string) *extensions.Ingress {
@@ -92,82 +86,6 @@ func TestConvertIngressControllerMode(t *testing.T) {
 		if ignore := class.IsValid(ing, ingressClass, defaultIngressClass); ignore != c.Ignore {
 			t.Errorf("convertIngressControllerMode(%q, %q), with Ingress annotation %q => "+
 				"Got ignore %v, want %v", c.Mode, "istio", c.Annotation, c.Ignore, ignore)
-		}
-	}
-}
-
-func TestSyncer(t *testing.T) {
-	client := fake.NewSimpleClientset()
-	ip := "64.233.191.200"
-	svc := v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "istio-ingress",
-			Namespace: namespace,
-		},
-		Status: v1.ServiceStatus{
-			LoadBalancer: v1.LoadBalancerStatus{
-				Ingress: []v1.LoadBalancerIngress{{
-					IP: ip,
-				}},
-			},
-		},
-	}
-	mesh := model.DefaultMeshConfig()
-	mesh.IngressService = svc.Name
-
-	if _, err := client.Core().Services(namespace).Create(&svc); err != nil {
-		t.Error(err)
-	}
-	if _, err := client.Extensions().Ingresses(namespace).Create(&ingress); err != nil {
-		t.Error(err)
-	}
-	pod := "test"
-	if _, err := client.Core().Pods(namespace).Create(&v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pod,
-			Namespace: namespace,
-		},
-		Spec: v1.PodSpec{
-			NodeName: "node",
-		},
-	}); err != nil {
-		t.Error(err)
-	}
-
-	// prior to invoking, need to set environment variables
-	// WARNING: this pollutes environment and it's recommended to run within a bazel sandbox
-	if err := os.Setenv("POD_NAME", pod); err != nil {
-		t.Error(err)
-	}
-	if err := os.Setenv("POD_NAMESPACE", namespace); err != nil {
-		t.Error(err)
-	}
-
-	stop := make(chan struct{})
-	if syncer, err := NewStatusSyncer(&mesh, client, namespace, kube.ControllerOptions{
-		WatchedNamespace: namespace,
-		ResyncPeriod:     resync,
-	}); err != nil {
-		t.Fatal(err)
-	} else {
-		go syncer.Run(stop)
-	}
-
-	count := 0
-	for {
-		if count > 60 {
-			t.Fatalf("did not set ingress status after one minute")
-		}
-		<-time.After(time.Second)
-		out, err := client.Extensions().Ingresses(namespace).Get(ingress.Name, metav1.GetOptions{})
-		if err != nil {
-			continue
-		}
-		ips := out.Status.LoadBalancer.Ingress
-		log.Infoa(ips)
-		if len(ips) > 0 && ips[0].IP == ip {
-			close(stop)
-			break
 		}
 	}
 }
