+++ tests/e2e/tests/pilot/ingress_controller_test.go
@@ -15,12 +15,15 @@
 package pilot
 
 import (
+	"os"
 	"reflect"
 	"testing"
 	"time"
 
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/kubernetes/fake"
 
@@ -30,6 +33,7 @@ import (
 	"istio.io/istio/pilot/pkg/serviceregistry/kube"
 	"istio.io/istio/pilot/test/mock"
 	"istio.io/istio/pilot/test/util"
+	"istio.io/istio/pkg/log"
 )
 
 const (
@@ -265,5 +269,81 @@ func TestIngressController(t *testing.T) {
 	}
 	if elts, err := ctl.List(model.IngressRule.Type, "missing"); err != nil || len(elts) > 0 {
 		t.Errorf("List() => got %#v, %v for a missing namespace", elts, err)
+	}
+}
+
+func TestSyncer(t *testing.T) {
+	client := fake.NewSimpleClientset()
+	ip := "64.233.191.200"
+	svc := v1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "istio-ingress",
+			Namespace: namespace,
+		},
+		Status: v1.ServiceStatus{
+			LoadBalancer: v1.LoadBalancerStatus{
+				Ingress: []v1.LoadBalancerIngress{{
+					IP: ip,
+				}},
+			},
+		},
+	}
+	mesh := model.DefaultMeshConfig()
+	mesh.IngressService = svc.Name
+
+	if _, err := client.Core().Services(namespace).Create(&svc); err != nil {
+		t.Error(err)
+	}
+	if _, err := client.Extensions().Ingresses(namespace).Create(&ig); err != nil {
+		t.Error(err)
+	}
+	pod := "test"
+	if _, err := client.Core().Pods(namespace).Create(&v1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      pod,
+			Namespace: namespace,
+		},
+		Spec: v1.PodSpec{
+			NodeName: "node",
+		},
+	}); err != nil {
+		t.Error(err)
+	}
+
+	// prior to invoking, need to set environment variables
+	// WARNING: this pollutes environment and it's recommended to run within a bazel sandbox
+	if err := os.Setenv("POD_NAME", pod); err != nil {
+		t.Error(err)
+	}
+	if err := os.Setenv("POD_NAMESPACE", namespace); err != nil {
+		t.Error(err)
+	}
+
+	stop := make(chan struct{})
+	if syncer, err := crd.NewStatusSyncer(&mesh, client, namespace, kube.ControllerOptions{
+		WatchedNamespace: namespace,
+		ResyncPeriod:     resync,
+	}); err != nil {
+		t.Fatal(err)
+	} else {
+		go syncer.Run(stop)
+	}
+
+	count := 0
+	for {
+		if count > 60 {
+			t.Fatalf("did not set ingress status after one minute")
+		}
+		<-time.After(time.Second)
+		out, err := client.Extensions().Ingresses(namespace).Get(ig.Name, metav1.GetOptions{})
+		if err != nil {
+			continue
+		}
+		ips := out.Status.LoadBalancer.Ingress
+		log.Infoa(ips)
+		if len(ips) > 0 && ips[0].IP == ip {
+			close(stop)
+			break
+		}
 	}
 }
